Move GetClusterThing out of pkg/utils.
utils isn't a useful package name, so moving this to pkg/network, which is a much better home for it.
