+++ pkg/reconciler/names.go
@@ -19,12 +19,12 @@ package reconciler
 import (
 	"fmt"
 
-	"github.com/knative/serving/pkg/utils"
+	"github.com/knative/serving/pkg/network"
 )
 
 // GetK8sServiceFullname returns service full name
 func GetK8sServiceFullname(name string, namespace string) string {
-	return fmt.Sprintf("%s.%s.svc.%s", name, namespace, utils.GetClusterDomainName())
+	return fmt.Sprintf("%s.%s.svc.%s", name, namespace, network.GetClusterDomainName())
 }
 
 // GetServingK8SServiceNameForObj returns the service name for the object
