+++ pkg/reconciler/v1alpha1/clusteringress/resources/virtual_service.go
@@ -32,9 +32,9 @@ import (
 	"github.com/knative/serving/pkg/apis/networking"
 	"github.com/knative/serving/pkg/apis/networking/v1alpha1"
 	"github.com/knative/serving/pkg/apis/serving"
+	"github.com/knative/serving/pkg/network"
 	"github.com/knative/serving/pkg/reconciler"
 	"github.com/knative/serving/pkg/reconciler/v1alpha1/clusteringress/resources/names"
-	"github.com/knative/serving/pkg/utils"
 )
 
 // MakeVirtualService creates an Istio VirtualService as network programming.
@@ -131,8 +131,8 @@ func expandedHosts(hosts []string) []string {
 	expanded := []string{}
 	allowedSuffixes := []string{
 		"",
-		"." + utils.GetClusterDomainName(),
-		".svc." + utils.GetClusterDomainName(),
+		"." + network.GetClusterDomainName(),
+		".svc." + network.GetClusterDomainName(),
 	}
 	for _, h := range hosts {
 		for _, suffix := range allowedSuffixes {
